Input acabado
