+++ PTSD.sln
@@ -8,8 +8,6 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PTSD-Core", "vsprojects\PTS
 		{BF836B04-6F79-4B76-BB79-C598B92D2DDA} = {BF836B04-6F79-4B76-BB79-C598B92D2DDA}
 	EndProjectSection
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PTSD-Input", "vsprojects\PTSD-Input\PTSDInput.vcxproj", "{6F50BAF9-27CC-4EB7-993A-ABC4AC9D9FA9}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PTSD-Physics", "vsprojects\PTSD-Physics\PTSD-Physics.vcxproj", "{04CF3F44-77C9-4B6B-B4E8-8A02268FBF45}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PTSD-Scripting", "vsprojects\PTSD-Scripting\PTSD-Scripting.vcxproj", "{999A09D0-B127-4481-926D-E215C121C364}"
@@ -28,10 +26,6 @@ Global
 		{CCE6C881-2767-4D8C-BA13-49A6C68C1E13}.Debug|x64.Build.0 = Debug|x64
 		{CCE6C881-2767-4D8C-BA13-49A6C68C1E13}.Release|x64.ActiveCfg = Release|x64
 		{CCE6C881-2767-4D8C-BA13-49A6C68C1E13}.Release|x64.Build.0 = Release|x64
-		{6F50BAF9-27CC-4EB7-993A-ABC4AC9D9FA9}.Debug|x64.ActiveCfg = Debug|x64
-		{6F50BAF9-27CC-4EB7-993A-ABC4AC9D9FA9}.Debug|x64.Build.0 = Debug|x64
-		{6F50BAF9-27CC-4EB7-993A-ABC4AC9D9FA9}.Release|x64.ActiveCfg = Release|x64
-		{6F50BAF9-27CC-4EB7-993A-ABC4AC9D9FA9}.Release|x64.Build.0 = Release|x64
 		{04CF3F44-77C9-4B6B-B4E8-8A02268FBF45}.Debug|x64.ActiveCfg = Debug|x64
 		{04CF3F44-77C9-4B6B-B4E8-8A02268FBF45}.Debug|x64.Build.0 = Debug|x64
 		{04CF3F44-77C9-4B6B-B4E8-8A02268FBF45}.Release|x64.ActiveCfg = Release|x64
