+++ source/Input/source/InputImplementation.h
@@ -9,7 +9,7 @@
 #include <queue>
 #include <map>
 #include <string>
-#include "Vec2.h"
+#include "PTSDVectors.h"
 
 namespace PTSD {
 	using namespace std;
